% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EScvtmle.R
\name{ES.cvtmle}
\alias{ES.cvtmle}
\title{ES.cvtmle}
\usage{
ES.cvtmle(
  txinrwd,
  data,
  study,
  covariates,
  treatment_var,
  treatment,
  outcome,
  NCO = NULL,
  Delta = NULL,
  Delta_NCO = NULL,
  pRCT,
  V = 10,
  Q.SL.library,
  d.SL.library,
  g.SL.library,
  Q.discreteSL,
  d.discreteSL,
  g.discreteSL,
  family,
  family_nco,
  fluctuation = "logistic",
  comparisons = list(c(1), c(1, 2)),
  adjustnco = FALSE,
  target.gwt = TRUE
)
}
\arguments{
\item{txinrwd}{Whether active treatment is available in RWD (TRUE/FALSE)}

\item{data}{The dataset}

\item{study}{Character name of variable indicating study participation (e.g. "S"). This variable should take values of 1 for the RCT and can take other values for the other study. Note that the code is currently set up only to handle two studies, but may be expanded to handle multiple studies in the future.}

\item{covariates}{Vector of character names of covariates to be adjusted for (e.g. c("W1", "W2"))}

\item{treatment_var}{Character name of treatment variable (e.g. "A")}

\item{treatment}{Name of treatment of interest (e.g. "DrugName" or 1)}

\item{outcome}{Character name of outcome variable (e.g. "Y"). If Y is a binary variable subject to censoring, it should be coded as 0 for observations that were censored.}

\item{NCO}{Character name of negative control outcome variable (e.g. "nco") or NULL if no NCO available. If NCO is a binary variable subject to censoring, it should be coded as 0 for observations that were censored.}

\item{Delta}{Character name of a variable that is 0 if an observation was censored (missing binary outcome) and 1 otherwise. Missing outcomes may also be coded as NA, in which case a Delta variable will be added internally. If no missing outcomes, set Delta=NULL.}

\item{Delta_NCO}{Character name of a variable that is 0 if the value of NCO is missing and 1 otherwise. Missing NCOs may also be coded as NA, in which case a Delta_NCO variable will be added internally. If no missing NCO or no NCO, set Delta_NCO=NULL.}

\item{pRCT}{The probability of randomization to treatment in the RCT}

\item{V}{Number of cross-validation folds (default 10).}

\item{Q.SL.library}{Candidate algorithms for SuperLearner estimation of outcome regressions}

\item{d.SL.library}{Candidate algorithms for SuperLearner estimation of missingness mechanism}

\item{g.SL.library}{Candidate algorithms for SuperLearner estimation of treatment mechanism for combined RCT/RWD analysis}

\item{Q.discreteSL}{Should a discrete SuperLearner be used for estimation of outcome regressions? (TRUE/FALSE)}

\item{d.discreteSL}{Should a discrete SuperLearner be used for estimation of missingness mechanism? (TRUE/FALSE)}

\item{g.discreteSL}{Should a discrete SuperLearner be used for estimation of treatment mechanism? (TRUE/FALSE)}

\item{family}{Either "binomial" for binary outcomes or "gaussian" for continuous outcomes}

\item{family_nco}{Family for negative control outcome}

\item{fluctuation}{'logistic' (default for binary and continuos outcomes), or 'linear' describing fluctuation for TMLE updating. If 'logistic' with a continuous outcome, outcomes are scaled to (0,1) for TMLE targeting and then returned to the original scale for parameter estimation.}

\item{comparisons}{A vector of the values of study variable S that you would like to consider. For example, if you have an RCT labeled S=1 and RWD labeled S=2, you would use comparisons = list(c(1),c(1,2)) to compare RCT only to RCT + RWD.}

\item{adjustnco}{Should we adjust for the NCO as a proxy of bias in the estimation of the ATE of A on Y? (TRUE/FALSE)}

\item{target.gwt}{As in tmle package, when TRUE, move g from denominator of clever covariate to the weight when fitting coefficient for TMLE updating.}
}
\value{
Returns the ATE estimate with 95% confidence intervals for the Experiment-Selector CV-TMLE and the proportion of folds in which RWD was included in the estimate.
}
\description{
Runs Experiment-Selector CV-TMLE for selecting and analyzing optimal experiment as RCT with or without RWD
}
\examples{
data(wash)
#For unbiased external controls, use:
dat <- wash[which(wash$study \%in\% c(1,2)),]
library(SuperLearner)
set.seed(2022)
results_rwd1 <- ES.cvtmle(txinrwd=TRUE,
                          data=dat, study="study",
                          covariates=c("aged", "sex", "momedu", "hfiacat"),
                          treatment_var="treatment", treatment=1,
                          outcome="laz", NCO="Nlt18scale",
                          Delta=NULL, Delta_NCO=NULL,
                          pRCT=0.5, V=10, Q.SL.library=c("SL.glm"),
                          g.SL.library=c("SL.glm"), Q.discreteSL=TRUE, g.discreteSL=TRUE,
                          family="gaussian", family_nco="gaussian", fluctuation = "logistic",
                          comparisons = list(c(1),c(1,2)), adjustnco = FALSE, target.gwt = TRUE)
print(results_rwd1)
}
