% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EScvtmle.R
\name{plot.EScvtmle}
\alias{plot.EScvtmle}
\title{plot.EScvtmle}
\usage{
\method{plot}{EScvtmle}(x, ...)
}
\arguments{
\item{x}{An object of class "EScvtmle"}

\item{...}{Other arguments to plot}
}
\description{
Plots fold-specific ATE estimates and a histogram of Monte Carlo sample ATE estimates used to construct confidence intervals.
}
